+++ src/lib/starter-templates/base-files/ensure-cspell.test.ts
@@ -8,6 +8,8 @@ describe('ensureCspell', () => {
   const expectedDefaultWords = [
     'bradlc',
     'bunx',
+    'coep',
+    'hsts',
     'dbaeumer',
     'esbenp',
     'eslintcache',